+++ lib/stsci/tools/tests/test_breadcrumb.py
@@ -1,6 +1,4 @@
 """This was pandokia breadcrumb test in stuff.py"""
-from __future__ import absolute_import
-
 import os
 
 from .. import capable