+++ lib/stsci/tools/xyinterp.py
@@ -7,17 +7,16 @@ Interpolates y based on the given xval.
 be the same length. The x array must also be sorted.
 `xval` is a user-specified value. This routine looks
 up `xval` in the x array and uses that information to properly interpolate
-the value in the y array. 
+the value in the y array.
 
 
 :author: Vicki Laidler
 
-:version: '0.1 (2006-07-06)'  
+:version: '0.1 (2006-07-06)'
 
 
 """
-from __future__ import division # confidence high
-import numpy as N 
+import numpy as N
 
 #This section for standalone imports only-------------------------------------
 __version__ = '0.1'          #Release version number only
@@ -26,47 +25,47 @@ __vdate__ = '2006-07-06'     #Date of this version, in this (FITS-style) format
 
 
 def xyinterp(x,y,xval):
-    """ 
-    
+    """
+
     :Purpose: Interpolates y based on the given xval.
 
     x and y are a pair of independent/dependent variable arrays that must
     be the same length. The x array must also be sorted.
     xval is a user-specified value. This routine looks
     up xval in the x array and uses that information to properly interpolate
-    the value in the y array.  
+    the value in the y array.
 
     Notes
     =====
     Use the searchsorted method on the X array to determine the bin in
     which xval falls; then use that information to compute the corresponding
     y value.
-    
 
-    See Also 
+
+    See Also
     ========
     numpy
 
     Parameters
     ==========
 
-    x: 1D numpy array  
+    x: 1D numpy array
         independent variable array: MUST BE SORTED
 
     y: 1D numpy array
         dependent variable array
 
-    xval: float 
+    xval: float
         the x value at which you want to know the value of y
 
     Returns
     =======
-    y: float 
+    y: float
         the value of y corresponding to xval
 
     Raises
     ======
-    ValueError: 
+    ValueError:
         If arrays are unequal length; or x array is unsorted;
         or if xval falls outside the bounds of x (extrapolation is unsupported
 
@@ -88,11 +87,11 @@ def xyinterp(x,y,xval):
     #This algorithm only works on sorted data
     if x.argsort().all() != N.arange(len(x)).all():
         raise ValueError("Input array x must be sorted")
-    
+
     # Now do the real work.
     hi = x.searchsorted(xval)
     lo = hi - 1
-    
+
     try:
         seg = (float(xval)-x[lo]) / (x[hi] - x[lo])
     except ZeroDivisionError: